# back/public/api-docs.yaml
openapi: 3.0.0
info:
  title: 'API ProjetPro'
  description: "API REST pour l'application ProjetPro avec système d'authentification sécurisé.\n\n## Fonctionnalités principales:\n- Authentification avec protection anti-force brute\n- Gestion des tokens CSRF\n- Upload d'images vers Azure Blob Storage\n- Logging automatique des tentatives de connexion\n- Système de blocage par IP et par login\n\n## Sécurité:\n- Protection CSRF obligatoire\n- Tokens d'authentification JWT\n- Blocage automatique après échecs répétés\n\n## Headers requis:\n- `X-CSRF-Token`: Token CSRF pour les requêtes POST\n- `Authorization: Bearer <token>`: Token JWT pour l'authentification\n- `Content-Type: application/json`\n"
  contact:
    name: 'Support API'
    email: support@projetpro.com
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8000'
    description: 'Serveur de développement'
  -
    url: /api
    description: 'API Production'
paths:
...
  /api/login:
    post:
      tags:
        - Authentication
      summary: 'Authentifier un utilisateur'
      description: 'Authentifie un utilisateur avec son login et mot de passe. Inclut une protection contre les attaques par force brute.'
      operationId: authenticateUser
      requestBody:
        description: 'Identifiants de connexion utilisateur'
        required: true
        content:
          application/json:
            schema:
              required:
                - login
                - password
              properties:
                login:
                  description: "Nom d'utilisateur ou adresse email"
                  type: string
                  example: john.doe
                password:
                  description: "Mot de passe de l'utilisateur"
                  type: string
                  format: password
                  example: motDePasseSecret123
              type: object
      responses:
        '200':
          description: 'Connexion réussie avec tokens JWT'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JWTLoginResponseDTO'
        '401':
          description: 'Identifiants incorrects'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
        '429':
          description: 'Trop de tentatives - Protection anti-brute force'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
      security:
        -
          csrfToken: []
    options:
      tags:
        - Authentication
      summary: "Options pour l'authentification utilisateur"
      description: "Retourne les options disponibles pour l'authentification utilisateur (CORS, headers, etc)."
      operationId: loginOptions
      responses:
        '200':
          description: 'Options retournées'
          content:
            application/json:
              schema:
                properties:
                  Access-Control-Allow-Origin: { type: string, example: '*' }
                  Access-Control-Allow-Methods: { type: string, example: 'POST, OPTIONS' }
                  Access-Control-Allow-Headers: { type: string, example: 'Content-Type, X-CSRF-Token' }
                type: object
  /api/register:
    post:
      tags:
        - Authentication
      summary: 'Inscrire un nouvel utilisateur'
      description: "Crée un nouvel utilisateur avec un login et un mot de passe. Vérifie la validité et l'unicité du login."
      operationId: registerUser
      requestBody:
        description: "Identifiants pour l'inscription"
        required: true
        content:
          application/json:
            schema:
              required:
                - username
                - password
              properties:
                username:
                  description: "Nom d'utilisateur"
                  type: string
                  example: john.doe
                password:
                  description: 'Mot de passe'
                  type: string
                  format: password
                  example: motDePasseSecret123
              type: object
      responses:
        '201':
          description: 'Utilisateur inscrit avec succès'
          content:
            application/json:
              schema:
                properties:
                  message: { type: string, example: 'User registered successfully' }
                type: object
        '400':
          description: 'Paramètres invalides ou manquants'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
        '409':
          description: "Nom d'utilisateur déjà existant"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
        '500':
          description: 'Erreur serveur'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
...
  /api/tokens/info:
    get:
      tags:
        - 'JWT Tokens'
      summary: "Obtenir les informations d'un token JWT"
      description: 'Retourne les informations détaillées du token JWT transmis via cookie ou header.'
      operationId: getTokenInfo
      responses:
        '200':
          description: 'Informations du token'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenInfoResponseDTO'
        '401':
          description: 'Token manquant ou invalide'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
        '500':
          description: 'Erreur serveur'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
...
components:
  schemas:
...
    ErrorResponseDTO:
      required:
        - error
      properties:
        error:
          type: string
        message:
          type: string
          default: null
          nullable: true
        code:
          type: integer
          default: null
          nullable: true
        details:
          type: array
          items: {  }
          default: null
          nullable: true
      type: object
...
    JWTLoginResponseDTO:
      required:
        - success
        - message
        - user
        - tokens
      properties:
        success:
          title: 'Crée un DTO de succès avec tokens JWT.'
          type: boolean
        message:
          type: string
        user:
          $ref: '#/components/schemas/LoginUserDTO'
        tokens:
          type: array
          items: {  }
      type: object
...
    TokenInfoResponseDTO:
      required:
        - valid
        - stillValid
      properties:
        valid:
          type: boolean
        tokenId:
          type: string
          default: null
          nullable: true
        userId:
          type: string
          default: null
          nullable: true
        username:
          type: string
          default: null
          nullable: true
        tokenType:
          type: string
          default: null
          nullable: true
        issuedAt:
          type: string
          nullable: true
        expiresAt:
          type: string
          default: null
          nullable: true
        roles:
          type: array
          items: {  }
          default: null
          nullable: true
        issuer:
          type: string
          nullable: true
        audience:
          type: string
          default: null
          nullable: true
        loginTime:
          type: integer
          default: null
          nullable: true
        ipAddress:
          type: string
          default: null
          nullable: true
        userAgent:
          type: string
          default: null
          nullable: true
        error:
          type: string
          default: null
          nullable: true
        stillValid:
          title: 'Vérifie si le token est encore valide (non expiré).'
          type: boolean
        timeUntilExpiration:
          title: 'Obtient le temps restant avant expiration en secondes.'
          type: integer
          nullable: true
      type: object
    LoginUserDTO:
      title: 'Informations utilisateur'
      description: "Données utilisateur dans les réponses d'authentification"
      required:
        - id
        - username
        - roles
      properties:
        id:
          description: "Identifiant unique de l'utilisateur"
          type: integer
          example: 1
        username:
          description: "Nom d'utilisateur"
          type: string
          example: john.doe
        roles:
          description: "Liste des rôles de l'utilisateur"
          type: array
          items:
            type: string
          example:
            - ROLE_USER
        last_visit:
          description: 'Date et heure de la dernière visite'
          type: string
          format: date-time
          default: null
          example: '2025-06-29 14:30:00'
          nullable: true
      type: object
...
  securitySchemes:
    bearerAuth:
      type: http
      description: "Token JWT d'authentification"
      bearerFormat: JWT
      scheme: bearer
    csrfToken:
      type: apiKey
      description: 'Token CSRF pour la protection contre les attaques CSRF'
      name: X-CSRF-Token
      in: header
security:
  -
    bearerAuth: []
  -
    csrfToken: []
...